requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0

# Technical Analysis
ta==0.10.2
pandas-ta==0.3.14b0

# Data Fetching
yfinance==0.2.28
requests==2.31.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Backtesting
backtrader==1.9.78

# Web App
streamlit==1.28.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0

# Development
jupyter==1.0.0
ipython==8.15.0
pytest==7.4.0
